add multicore functionality - runs indipendent program
